requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
flask-cors==4.0.1